+++ Changelog.md
@@ -1,10 +1,11 @@
 ## 0.0.51
 + Added `muna.beta.predictions.remote.stream` method to stream remote predictions.
-+ Added support for `aac`, `flac`, `opus`, `wav` response formats in `muna.beta.openai.audio.speech.create` function.
++ Added support for `aac`, `flac`, `mp3`, `opus`, `wav` response formats in `muna.beta.openai.audio.speech.create` function.
 + Refactored `Acceleration` constants to use a `local_` prefix.
 + Refactored `Parameter.type` field to `Parameter.dtype`.
++ Replaced `Parameter.range` field with `Parameter.min` and `Parameter.max` fields.
 + Removed `ParameterDenotation` type.
-+ Upgraded to Function C 0.0.39.
++ Upgraded to Function C 0.0.40.
 
 ## 0.0.50
 + Added `muna.beta.openai.embeddings.create` method for using text embedding predictors via the OpenAI client.